requirements.txt
fastapi
uvicorn
python-multipart
openai-whisper
python-docx
reportlab
torch
setuptools-rust
